requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.6.0
Django==4.1.1
django-crispy-forms==1.14.0
django-htmx==1.12.2
fonttools==4.37.1
html5lib==1.1
Pillow==9.2.0
pycparser==2.21
pydyf==0.2.0
pyphen==0.13.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
weasyprint==56.1
webencodings==0.5.1
zopfli==0.2.1
